Performance fixes needed for production
These changes were needed to successfully process large
databases in a frequently updated cBioPortal server:
- instead of iterating through all genetic profiles one by one,
    sometimes failing for profiles with inconsistant event list
    lengths, the problematic genetic profiles are identified
    first and excluded during an "async" mode clickhouse command
    which operates (in one operation) across all other profiles
- sling operations for genetic_alteration can be split into 3
    chunks by setting an environment variable in order work
    around timeouts and disconnect issues
- verification of complete cloning and copying of the
    genetic_alteration table (via record count comparison)
    can be inhibited by setting an environment variable
